fix(seo-audit): add column whitelist in Page::upsert() to prevent DB errors
CrawlerService::crawlPage() returns extra fields (error, links, headings,
images, budget) that don't exist in sa_pages table. Added VALID_COLUMNS
filtering in Page::upsert() so invalid columns never reach Database::update().
Also fixed float-to-int deprecation warning in CrawlerService::rateLimit().

+++ modules/seo-audit/services/CrawlerService.php
@@ -965,7 +965,7 @@ class CrawlerService
         $elapsed = $now - $this->lastRequestTime;
 
         if ($elapsed < $this->requestDelay) {
-            usleep(($this->requestDelay - $elapsed) * 1000);
+            usleep((int) (($this->requestDelay - $elapsed) * 1000));
         }
 
         $this->lastRequestTime = microtime(true) * 1000;
